Created a LangsBaseCommand.
